Extract entry icon component
