+++ src/shared/components/results/ResultsContainer.tsx
@@ -39,26 +39,12 @@ const defaultTableColumns: Partial<Record<Namespace, AllColumns>> = {
     UniProtKBColumn.organismName,
   ],
   [Namespace.uniref]: [
-    // UniRefColumn.id,
-    // UniRefColumn.name,
-    // UniRefColumn.count,
-    // UniRefColumn.organism,
-    // UniRefColumn.length,
-    // UniRefColumn.identity,
-
     UniRefColumn.id,
     UniRefColumn.name,
-    UniRefColumn.commonTaxon,
-    UniRefColumn.commonTaxonid,
-    UniRefColumn.organismId,
-    UniRefColumn.organism,
-    UniRefColumn.identity,
-    UniRefColumn.length,
-    UniRefColumn.sequence,
-    UniRefColumn.types,
-    UniRefColumn.members,
     UniRefColumn.count,
-    UniRefColumn.created,
+    UniRefColumn.organism,
+    UniRefColumn.length,
+    UniRefColumn.identity,
   ],
 };
