+++ src/uniref/config/UniRefColumnConfiguration.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable camelcase */
 import React from 'react';
 import { Link } from 'react-router-dom';
-import { EntryTypeIcon } from '../../shared/components/entry/EntryTitle';
+import EntryTypeIcon from '../../shared/components/entry/EntryTypeIcon';
 import { UniRefLiteAPIModel } from '../adapters/uniRefConverter';
 
 export enum UniRefColumn {
